move methods from api
