+++ contracts/satoshi-bridge/src/api/bridge.rs
@@ -1,4 +1,3 @@
-use crate::psbt_wrapper::PsbtWrapper;
 use crate::*;
 use near_plugins::{access_control_any, pause};
 
@@ -449,249 +448,5 @@ impl Contract {
     pub fn get_change_address(&self) -> Option<String> {
         let config = self.internal_config();
         config.change_address.clone()
-    }
-}
-
-impl Contract {
-    pub fn create_active_utxo_management_pending_info(
-        &mut self,
-        account_id: AccountId,
-        mut psbt: PsbtWrapper,
-    ) {
-        let account = self.internal_unwrap_account(&account_id);
-        require!(
-            account.btc_pending_sign_id.is_none(),
-            "Previous btc tx has not been signed"
-        );
-
-        let (utxo_storage_keys, vutxos) = self.generate_vutxos(&mut psbt);
-        let (actual_received_amount, gas_fee) =
-            self.check_active_management_psbt_valid(&psbt, &vutxos);
-        require!(
-            gas_fee <= self.data().cur_available_protocol_fee,
-            "Insufficient protocol_fee"
-        );
-        self.data_mut().cur_available_protocol_fee -= gas_fee;
-        self.data_mut().cur_reserved_protocol_fee += gas_fee;
-
-        let need_signature_num = psbt.get_input_num();
-        let psbt_hex = psbt.serialize();
-        let btc_pending_id = psbt.get_pending_id();
-        let btc_pending_info = BTCPendingInfo {
-            account_id: account_id.clone(),
-            btc_pending_id: btc_pending_id.clone(),
-            transfer_amount: 0,
-            actual_received_amount,
-            withdraw_fee: 0,
-            gas_fee,
-            burn_amount: gas_fee,
-            psbt_hex,
-            vutxos,
-            signatures: vec![None; need_signature_num],
-            tx_bytes_with_sign: None,
-            create_time_sec: nano_to_sec(env::block_timestamp()),
-            last_sign_time_sec: 0,
-            state: PendingInfoState::ActiveUtxoManagementOriginal(OriginalState {
-                stage: PendingInfoStage::PendingSign,
-                max_gas_fee: gas_fee,
-                last_rbf_time_sec: None,
-                cancel_rbf_reserved: None,
-            }),
-        };
-        require!(
-            self.data_mut()
-                .btc_pending_infos
-                .insert(btc_pending_id.clone(), btc_pending_info.into())
-                .is_none(),
-            "pending info already exist"
-        );
-        self.internal_unwrap_mut_account(&account_id)
-            .btc_pending_sign_id = Some(btc_pending_id.clone());
-        Event::UtxoRemoved { utxo_storage_keys }.emit();
-        Event::GenerateBtcPendingInfo {
-            account_id: &account_id,
-            btc_pending_id: &btc_pending_id,
-        }
-        .emit();
-    }
-
-    fn internal_verify_deposit_entry(
-        &mut self,
-        deposit_msg: DepositMsg,
-        tx_bytes: Vec<u8>,
-        vout: usize,
-        tx_block_blockhash: String,
-        tx_index: u64,
-        merkle_proof: Vec<String>,
-        coinbase_proof: Option<(String, Vec<String>)>,
-    ) -> Promise {
-        require!(
-            deposit_msg.safe_deposit.is_none(),
-            "safe_deposit not supported in verify_deposit"
-        );
-        let path = get_deposit_path(&deposit_msg);
-        let transaction = WrappedTransaction::decode(&tx_bytes, &self.internal_config().chain)
-            .expect("Deserialization tx_bytes failed");
-        let deposit_amount = u128::from(transaction.output()[vout].value.to_sat());
-        require!(deposit_amount > 0, "Invalid deposit_amount");
-        require!(
-            transaction.lock_time() == LockTime::ZERO,
-            "Tx with a non-zero lock_time are not supported."
-        );
-        let deposit_address = self.generate_utxo_chain_address(&path);
-        let deposit_address_script_pubkey = deposit_address
-            .script_pubkey()
-            .expect("Invalid deposit address");
-        require!(
-            deposit_address_script_pubkey == transaction.output()[vout].script_pubkey,
-            "Invalid deposit tx_bytes"
-        );
-
-        let utxo = UTXO {
-            path,
-            tx_bytes,
-            vout,
-            balance: transaction.output()[vout].value.to_sat(),
-        };
-        let tx_id = transaction.compute_txid().to_string();
-        let utxo_storage_key = generate_utxo_storage_key(
-            tx_id.clone(),
-            u32::try_from(vout).unwrap_or_else(|_| env::panic_str("vout overflow")),
-        );
-        self.internal_verify_deposit(
-            deposit_amount,
-            tx_block_blockhash,
-            tx_index,
-            merkle_proof,
-            coinbase_proof,
-            PendingUTXOInfo {
-                tx_id,
-                utxo_storage_key,
-                utxo,
-            },
-            deposit_msg,
-        )
-    }
-
-    fn internal_safe_verify_deposit_entry(
-        &mut self,
-        deposit_msg: DepositMsg,
-        tx_bytes: Vec<u8>,
-        vout: usize,
-        tx_block_blockhash: String,
-        tx_index: u64,
-        merkle_proof: Vec<String>,
-        coinbase_proof: Option<(String, Vec<String>)>,
-    ) -> Promise {
-        require!(
-            env::attached_deposit() >= self.required_balance_for_safe_deposit(),
-            "Insufficient deposit for storage"
-        );
-
-        let path = get_deposit_path(&deposit_msg);
-        let safe_deposit_msg = deposit_msg
-            .safe_deposit
-            .unwrap_or_else(|| env::panic_str("safe_deposit is required in safe_verify_deposit"));
-
-        let transaction = WrappedTransaction::decode(&tx_bytes, &self.internal_config().chain)
-            .expect("Deserialization tx_bytes failed");
-        let deposit_amount = transaction.output()[vout].value.to_sat().into();
-        require!(deposit_amount > 0, "Invalid deposit_amount");
-        require!(
-            transaction.lock_time() == LockTime::ZERO,
-            "Tx with a non-zero lock_time are not supported."
-        );
-        let deposit_address = self.generate_utxo_chain_address(&path);
-        let deposit_address_script_pubkey = deposit_address
-            .script_pubkey()
-            .expect("Invalid deposit address");
-        require!(
-            deposit_address_script_pubkey == transaction.output()[vout].script_pubkey,
-            "Invalid deposit tx_bytes"
-        );
-
-        let utxo = UTXO {
-            path,
-            tx_bytes,
-            vout,
-            balance: transaction.output()[vout].value.to_sat(),
-        };
-        let tx_id = transaction.compute_txid().to_string();
-        let utxo_storage_key = generate_utxo_storage_key(tx_id.clone(), vout.try_into().unwrap());
-
-        self.internal_safe_verify_deposit(
-            deposit_amount,
-            tx_block_blockhash,
-            tx_index,
-            merkle_proof,
-            coinbase_proof,
-            PendingUTXOInfo {
-                tx_id,
-                utxo_storage_key,
-                utxo,
-            },
-            deposit_msg.recipient_id,
-            safe_deposit_msg,
-        )
-    }
-
-    fn internal_verify_withdraw_entry(
-        &mut self,
-        tx_id: String,
-        tx_block_blockhash: String,
-        tx_index: u64,
-        merkle_proof: Vec<String>,
-        coinbase_proof: Option<(String, Vec<String>)>,
-    ) -> Promise {
-        let btc_pending_info = self.internal_unwrap_btc_pending_info(&tx_id);
-        btc_pending_info.assert_withdraw_related_pending_verify_tx();
-        if let Some(original_tx_id) = btc_pending_info.get_original_tx_id() {
-            require!(
-                self.check_btc_pending_info_exists(original_tx_id),
-                "original tx already verified"
-            );
-        }
-        require!(
-            btc_pending_info.tx_bytes_with_sign.is_some(),
-            "Missing tx_bytes_with_sign"
-        );
-        self.internal_verify_withdraw(
-            tx_id,
-            tx_block_blockhash,
-            tx_index,
-            merkle_proof,
-            coinbase_proof,
-            btc_pending_info,
-        )
-    }
-
-    fn internal_verify_active_utxo_management_entry(
-        &mut self,
-        tx_id: String,
-        tx_block_blockhash: String,
-        tx_index: u64,
-        merkle_proof: Vec<String>,
-        coinbase_proof: Option<(String, Vec<String>)>,
-    ) -> Promise {
-        let btc_pending_info = self.internal_unwrap_btc_pending_info(&tx_id);
-        btc_pending_info.assert_active_utxo_management_related_pending_verify_tx();
-        if let Some(original_tx_id) = btc_pending_info.get_original_tx_id() {
-            require!(
-                self.check_btc_pending_info_exists(original_tx_id),
-                "original tx already verified"
-            );
-        }
-        require!(
-            btc_pending_info.tx_bytes_with_sign.is_some(),
-            "Missing tx_bytes_with_sign"
-        );
-        self.internal_verify_active_utxo_management(
-            tx_id,
-            tx_block_blockhash,
-            tx_index,
-            merkle_proof,
-            coinbase_proof,
-            btc_pending_info,
-        )
     }
 }
