+++ contracts/satoshi-bridge/src/btc_light_client/deposit.rs
@@ -2,11 +2,12 @@ use near_sdk::serde_json::Value;
 
 use crate::{
     burn::GAS_FOR_BURN_CALL,
-    env, ext_nbtc,
+    deposit_msg::get_deposit_path,
+    env, ext_nbtc, generate_utxo_storage_key,
     mint::{GAS_FOR_MINT_CALL, GAS_FOR_MINT_CALL_BACK},
-    near, require, serde_json, AccountId, Contract, ContractExt, DepositMsg, Event, Gas, NearToken,
-    PendingUTXOInfo, PostAction, Promise, PromiseOrValue, SafeDepositMsg, MAX_BOOL_RESULT,
-    MAX_FT_TRANSFER_CALL_RESULT, U128,
+    near, require, serde_json, AccountId, Contract, ContractExt, DepositMsg, Event, Gas, LockTime,
+    NearToken, PendingUTXOInfo, PostAction, Promise, PromiseOrValue, SafeDepositMsg,
+    WrappedTransaction, MAX_BOOL_RESULT, MAX_FT_TRANSFER_CALL_RESULT, U128, UTXO,
 };
 
 pub const GAS_FOR_VERIFY_DEPOSIT_CALL_BACK: Gas = Gas::from_tgas(190);
@@ -110,6 +111,126 @@ impl Contract {
                     ),
             )
         }
+    }
+
+    pub(crate) fn internal_verify_deposit_entry(
+        &mut self,
+        deposit_msg: DepositMsg,
+        tx_bytes: Vec<u8>,
+        vout: usize,
+        tx_block_blockhash: String,
+        tx_index: u64,
+        merkle_proof: Vec<String>,
+        coinbase_proof: Option<(String, Vec<String>)>,
+    ) -> Promise {
+        require!(
+            deposit_msg.safe_deposit.is_none(),
+            "safe_deposit not supported in verify_deposit"
+        );
+        let path = get_deposit_path(&deposit_msg);
+        let transaction = WrappedTransaction::decode(&tx_bytes, &self.internal_config().chain)
+            .expect("Deserialization tx_bytes failed");
+        let deposit_amount = u128::from(transaction.output()[vout].value.to_sat());
+        require!(deposit_amount > 0, "Invalid deposit_amount");
+        require!(
+            transaction.lock_time() == LockTime::ZERO,
+            "Tx with a non-zero lock_time are not supported."
+        );
+        let deposit_address = self.generate_utxo_chain_address(&path);
+        let deposit_address_script_pubkey = deposit_address
+            .script_pubkey()
+            .expect("Invalid deposit address");
+        require!(
+            deposit_address_script_pubkey == transaction.output()[vout].script_pubkey,
+            "Invalid deposit tx_bytes"
+        );
+
+        let utxo = UTXO {
+            path,
+            tx_bytes,
+            vout,
+            balance: transaction.output()[vout].value.to_sat(),
+        };
+        let tx_id = transaction.compute_txid().to_string();
+        let utxo_storage_key = generate_utxo_storage_key(
+            tx_id.clone(),
+            u32::try_from(vout).unwrap_or_else(|_| env::panic_str("vout overflow")),
+        );
+        self.internal_verify_deposit(
+            deposit_amount,
+            tx_block_blockhash,
+            tx_index,
+            merkle_proof,
+            coinbase_proof,
+            PendingUTXOInfo {
+                tx_id,
+                utxo_storage_key,
+                utxo,
+            },
+            deposit_msg,
+        )
+    }
+
+    pub(crate) fn internal_safe_verify_deposit_entry(
+        &mut self,
+        deposit_msg: DepositMsg,
+        tx_bytes: Vec<u8>,
+        vout: usize,
+        tx_block_blockhash: String,
+        tx_index: u64,
+        merkle_proof: Vec<String>,
+        coinbase_proof: Option<(String, Vec<String>)>,
+    ) -> Promise {
+        require!(
+            env::attached_deposit() >= self.required_balance_for_safe_deposit(),
+            "Insufficient deposit for storage"
+        );
+
+        let path = get_deposit_path(&deposit_msg);
+        let safe_deposit_msg = deposit_msg
+            .safe_deposit
+            .unwrap_or_else(|| env::panic_str("safe_deposit is required in safe_verify_deposit"));
+
+        let transaction = WrappedTransaction::decode(&tx_bytes, &self.internal_config().chain)
+            .expect("Deserialization tx_bytes failed");
+        let deposit_amount = transaction.output()[vout].value.to_sat().into();
+        require!(deposit_amount > 0, "Invalid deposit_amount");
+        require!(
+            transaction.lock_time() == LockTime::ZERO,
+            "Tx with a non-zero lock_time are not supported."
+        );
+        let deposit_address = self.generate_utxo_chain_address(&path);
+        let deposit_address_script_pubkey = deposit_address
+            .script_pubkey()
+            .expect("Invalid deposit address");
+        require!(
+            deposit_address_script_pubkey == transaction.output()[vout].script_pubkey,
+            "Invalid deposit tx_bytes"
+        );
+
+        let utxo = UTXO {
+            path,
+            tx_bytes,
+            vout,
+            balance: transaction.output()[vout].value.to_sat(),
+        };
+        let tx_id = transaction.compute_txid().to_string();
+        let utxo_storage_key = generate_utxo_storage_key(tx_id.clone(), vout.try_into().unwrap());
+
+        self.internal_safe_verify_deposit(
+            deposit_amount,
+            tx_block_blockhash,
+            tx_index,
+            merkle_proof,
+            coinbase_proof,
+            PendingUTXOInfo {
+                tx_id,
+                utxo_storage_key,
+                utxo,
+            },
+            deposit_msg.recipient_id,
+            safe_deposit_msg,
+        )
     }
 }
